rename scorers_integrations to scorers_tests, fix scorers imports in tests
