+++ pyproject.toml
@@ -67,7 +67,7 @@ llamaindex = ["llama-index>=0.10.35"]
 mistral0 = ["mistralai>=0.1.8,<1.0.0"]
 mistral1 = ["mistralai>=1.0.0"]
 scorers = ["Levenshtein>=0.26.0", "instructor>=1.5.2"]
-scorers_integrations = ["openai>=1.0.0", "instructor>=1.5.2", "Levenshtein>=0.26.0", "google-generativeai>=0.8.0", "mistralai>=1.0.3", "anthropic>=0.30.0"]
+scorers_tests = ["instructor>=1.5.2", "Levenshtein>=0.26.0", "openai>=1.0.0", "google-generativeai>=0.8.0", "mistralai>=1.0.3", "anthropic>=0.30.0"]
 notdiamond = ["notdiamond>=0.3.21", "litellm<=1.49.1"]
 openai = ["openai>=1.0.0"]
 modal = ["modal", "python-dotenv"]
