+++ weave/scorers/hallucination_scorer.py
@@ -3,9 +3,9 @@ from typing import List
 from pydantic import BaseModel, Field
 
 import weave
-from weave.flow.scorers.llm_scorer import InstructorLLMScorer
-from weave.flow.scorers.llm_utils import OPENAI_DEFAULT_MODEL, create
-from weave.flow.scorers.utils import stringify
+from weave.scorers.llm_scorer import InstructorLLMScorer
+from weave.scorers.llm_utils import OPENAI_DEFAULT_MODEL, create
+from weave.scorers.utils import stringify
 
 DEFAULT_HALLUCINATION_SYSTEM_PROMPT = """
 Given some <input_data> from a user and an <output> generated by an AI system, \
